backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.3
openai==1.57.4
pgvector==0.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
